+++ src/server/routes/index.ts
@@ -1,6 +1,7 @@
 import ModelsRoute from './models.route'
 import TokensRoute from './tokens.route'
 import OpenAIProxyRoute from './proxy.route'
+import AnthropicProxyRoute from './anthropic-proxy.route'
 
 const routes = [
   {
@@ -14,6 +15,10 @@ const routes = [
   {
     path: '/openai',
     handler: OpenAIProxyRoute,
+  },
+  {
+    path: '/anthropic/openai',
+    handler: AnthropicProxyRoute,
   },
 ]
